fix: move USD import check to check.py module
